Fix folder structure (#96)
* Generate Files Pagetab

Generate TSV page tab to add the files to a submission

* Imaging Projects

- Adjust stats export job to be project based instead of acc no based
- Add new query to retrieve all the data in a single hit to the database
- Delete unnecessary code
- Create script to generate page tab for a set of files

* code review changes

* fixed issue

* code review changes

+++ Exporter/src/main/java/uk/ac/ebi/biostd/exporter/jobs/stats/process/StatsService.java
@@ -19,7 +19,7 @@ public class StatsService {
         return SubStats.builder()
                 .accNo(accNo)
                 .imaging(submission.isImagingSubmission())
-                .files(submission.getFilesCount())
+                .files((int) submission.getFilesCount())
                 .filesSize(submission.getFilesSize())
                 .subFileSize(getFileSize(accNo, submission.getRelPath()))
                 .build();
